admin_portal_flask/requirements.txt
flask
flask-admin
flask-login
flask-sqlalchemy
